+++ ChocAn.DataCenterConsole/Models/ProviderDetailsViewModel.cs
@@ -34,12 +34,12 @@ namespace ChocAn.DataCenterConsole.Models
 {
     public class ProviderDetailsViewModel
     {
-        public decimal Id { get; set; }
+        public int Id { get; set; }
         public string Name { get; set; }
         public string Email { get; set; }
         public string StreetAddress { get; set; }
         public string City { get; set; }
         public string State { get; set; }
-        public decimal ZipCode { get; set; }
+        public int ZipCode { get; set; }
     }
 }
