+++ ChocAn.MemberService/Member.cs
@@ -40,13 +40,13 @@ namespace ChocAn.MemberRepository
     public class Member
     {
         [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
-        public decimal Id { get; set; }
+        public int Id { get; set; }
         public string Name { get; set; }
         public string Email { get; set; }
         public string StreetAddress { get; set; }
         public string City { get; set; }
         public string State { get; set; }
-        public decimal ZipCode { get; set; }
+        public int ZipCode { get; set; }
         public string Status { get; set; }
     }
 }
